+++ src/app/models/enum/classification.enum.ts
@@ -3,3 +3,11 @@ export enum classificationStatusesEnum {
   PROCESSING,
   ENABLED
 }
+
+export enum bannerClassificationStatusEnum {
+  NEW = 0,
+  IN_PROGRESS = 1,
+  ERROR = 2,
+  SUCCESS = 3,
+  FAILURE = 4,
+}