+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.5] - 2020-01-30
+### Added
+- Classification info
+### Changed
+- Layout of site's codes
+### Fixed
+- Loading publisher's stats
+
+## [1.5.4] - 2020-01-27
+### Fixed
+- Loading sites by date filter
+
 ## [1.5.4] - 2020-01-27
 ### Fixed
 - Loading sites by date filter
@@ -213,7 +225,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
 
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.5.4...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.5.5...develop
+[1.5.5]: https://github.com/adshares/adpanel/compare/v1.5.4...v1.5.5
 [1.5.4]: https://github.com/adshares/adpanel/compare/v1.5.3...v1.5.4
 [1.5.3]: https://github.com/adshares/adpanel/compare/v1.5.2...v1.5.3
 [1.5.2]: https://github.com/adshares/adpanel/compare/v1.5.1...v1.5.2